Correction of variables type

+++ src/Main.java
@@ -2,13 +2,13 @@ public class Main {
     public static void main(String[] args) {
         CreditPaymentService creditPaymentService = new CreditPaymentService();
 
-        double monthlyPaymentCreditTerm_1Year = (int) (creditPaymentService.calculate(1_000_000, 12));
+        int monthlyPaymentCreditTerm_1Year = (int) (creditPaymentService.calculate(1_000_000, 12));
         System.out.println(monthlyPaymentCreditTerm_1Year);
 
-        double monthlyPaymentCreditTerm_2Year = (int) (creditPaymentService.calculate(1_000_000, 24));
+        int monthlyPaymentCreditTerm_2Year = (int) (creditPaymentService.calculate(1_000_000, 24));
         System.out.println(monthlyPaymentCreditTerm_2Year);
 
-        double monthlyPaymentCreditTerm_3Year = (int) (creditPaymentService.calculate(1_000_000, 36));
+        int monthlyPaymentCreditTerm_3Year = (int) (creditPaymentService.calculate(1_000_000, 36));
         System.out.println(monthlyPaymentCreditTerm_3Year);
 
     }
